backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-core
langchain-google-genai
google-generativeai
python-dotenv
SpeechRecognition
python-multipart